backend/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
kafka-python==2.0.2
mysql-connector-python==8.0.30
flask-pymongo
python-dotenv
